app/requirements.txt
streamlit==1.28.0
pandas==2.1.1
numpy==1.24.3
plotly==5.17.0
pymongo==4.5.0
python-dotenv==1.0.0
joblib==1.3.2
scikit-learn==1.3.2
prophet==1.1.4